Increase sizeof delta

+++ src/absoluteencoder.cpp
@@ -62,7 +62,7 @@ AbsoluteEncoder::update()
         newPosition ^= gray;
     }
 
-    int     newDelta( (newPosition - position_) & (cpr_ - 1) );
+    long    newDelta( (newPosition - position_) & (cpr_ - 1) );
 
     if ( newDelta & (cpr_ >> 1) )
     {
